+++ src/refactoring/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { CartPage } from './components/CartPage.tsx';
-import { AdminPage } from './components/AdminPage.tsx';
+import { AdminPage } from './components/admin/AdminPage.tsx';
 import { Coupon, Product } from '../types.ts';
 import { useCoupons, useProducts } from "./hooks/cart/index.ts";
