+++ src/refactoring/hooks/admin/index.ts
@@ -1,3 +1,2 @@
 export * from "./useDiscount.ts";
-export * from "./useProductAccordion.ts";
 export * from "./useProductMange.ts";
