--- src/refactoring/hooks/admin/useProductAccordion.ts
@@ -1,19 +1,0 @@
-import { useState } from 'react';
-
-export const useProductAccordion = () => {
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds(prev => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  return { openProductIds, toggleProductAccordion };
-};
